+++ src/ItemSpawnerEnhanced/Core/VanillaItemCategories.cs
@@ -7,52 +7,57 @@ internal static class VanillaItemCategories
 {
     private static readonly IReadOnlyDictionary<string, ItemFilterTag> Tags = BuildTags();
 
-    public static ItemFilterTag Resolve(string displayName, string prefabName)
-    {
-        if (!Tags.TryGetValue(displayName, out ItemFilterTag tags))
-            Tags.TryGetValue(prefabName, out tags);
+    public static IReadOnlyDictionary<string, ItemFilterTag> ReviewedTags => Tags;
 
-        return tags;
-    }
+    public static ItemFilterTag Resolve(string prefabName) =>
+        Tags.TryGetValue(prefabName, out ItemFilterTag tags) ? tags : ItemFilterTag.None;
 
     private static IReadOnlyDictionary<string, ItemFilterTag> BuildTags()
     {
-        var tags = new Dictionary<string, ItemFilterTag>(StringComparer.OrdinalIgnoreCase);
+        var tags = new Dictionary<string, ItemFilterTag>(StringComparer.Ordinal);
         Add(tags, ItemFilterTag.Food,
-            "Airline Food", "Beehive", "Big Egg", "Big Lollipop", "Black Clusterberry",
-            "Blue Berrynana", "Blue Shroomberry", "Brown Berrynana", "Bugle Shroom",
-            "Button Shroom", "Chubby Shroom", "Cluster Shroom", "Coconut", "Cooked Bird",
-            "Egg", "Energy Drink", "Fortified Milk", "Frog Legs", "Gold Prickleberry",
-            "Granola Bar", "Green Crispberry", "Green Kingberry", "Green Shroomberry",
-            "Coconut Half", "Half-Coconut", "Honeycomb", "Hot Dog", "Mandrake",
-            "Marshmallow", "Medicinal Root", "Napberry", "Orange Winterberry",
-            "Pink Berrynana", "Purple Kingberry", "Purple Shroomberry", "Red Clusterberry",
-            "Red Crispberry", "Red Prickleberry", "Red Shroomberry", "Scorchberry",
-            "Scorpion", "Scout Cookies", "Small Egg", "Sports Drink", "The Early Worm",
-            "Tick", "Trail Mix", "Yellow Berrynana", "Yellow Clusterberry",
-            "Yellow Crispberry", "Yellow Kingberry", "Yellow Shroomberry",
-            "Yellow Winterberry");
+            "Airplane Food", "Beehive", "NestEgg", "Lollipop", "Lollipop_Prop",
+            "Clusterberry Black", "Berrynana Blue", "Shroomberry_Blue", "Berrynana Brown",
+            "Mushroom Lace", "Mushroom Lace Poison", "Mushroom Normie Poison",
+            "Mushroom Normie", "Mushroom Chubby", "Mushroom Cluster Poison",
+            "Mushroom Cluster", "Item_Coconut", "EggTurkey", "EggRaven", "Egg",
+            "Energy Drink", "FortifiedMilk", "FrogLegs", "Prickleberry_Gold", "Granola Bar",
+            "Apple Berry Green", "C_Pawn_f Variant", "C_Bishop_m Variant", "C_Rook_m",
+            "Kingberry Green", "C_Pawn_m", "C_Pawn_f", "C_Bishop_f", "C_Knight",
+            "C_Bishop_m", "C_King", "C_Queen", "C_Rook_f", "C_Rook_f Variant",
+            "Shroomberry_Green", "Item_Coconut_half", "Item_Honeycomb",
+            "Glizzy_CattailVariant", "Glizzy", "Mandrake", "Marshmallow",
+            "Mandrake_Hidden", "MedicinalRoot", "Napberry", "Winterberry Orange",
+            "Berrynana Pink", "Kingberry Purple", "Shroomberry_Purple", "Clusterberry Red",
+            "Apple Berry Red", "Prickleberry_Red", "Shroomberry_Red", "Pepper Berry",
+            "Scorpion", "ScoutCookies", "ScoutCookies_Vanilla", "NestEgg_Raven",
+            "Sports Drink", "EarlyWorm", "Bugfix", "TrailMix", "Berrynana Yellow",
+            "Clusterberry Yellow", "Apple Berry Yellow", "Kingberry Yellow",
+            "Shroomberry_Yellow", "Winterberry Yellow");
         Add(tags, ItemFilterTag.Consumable,
-            "Aloe Vera", "Antidote", "Anti-Zooka", "Balloon", "Balloon Bunch", "Bandages",
-            "Blowgun", "Bugle of Friendship", "Candlestick", "Cure-All", "Cursed Skull",
-            "Dynamite", "Faerie Lantern", "First Aid Kit", "Flare", "Heat Pack", "Lantern",
-            "Pandora's Lunchbox", "Parachute", "Remedy Fungus", "Rescue Claw",
-            "Ritual Dagger", "Scout Effigy", "Scoutmaster's Bugle", "Sunscreen",
-            "The Book of Bones", "Torch", "Warp Compass", "Warp Fungus");
+            "AloeVera", "Antidote", "AntiZooka", "Balloon", "BalloonBunch", "Bandages",
+            "HealingDart Variant", "Bugle_Magic", "Candle", "Cure-All", "Cursed Skull",
+            "Dynamite", "Lantern_Faerie", "FirstAidKit", "Flare", "Heat Pack", "Lantern",
+            "PandorasBox", "Parachute", "HealingPuffShroom", "RescueHook",
+            "RescueHook_Infinite", "RitualDagger", "ScoutEffigy", "Bugle_Scoutmaster Variant",
+            "Sunscreen", "BookOfBones", "Torch", "Warp Compass", "Cheat Compass",
+            "Cheat Compass 1", "WarpFungus");
         Add(tags, ItemFilterTag.Equipment,
-            "Ancient Idol", "Backpack", "Balloon", "Balloon Bunch", "Binoculars",
-            "Bugle", "Bugle?", "Compass", "Fanny Pack", "Fannypack", "Glider",
-            "Jetpack", "Parasol", "Parachute", "Pirate's Compass", "Rocketpack");
+            "AncientIdol", "Backpack", "Balloon", "BalloonBunch", "Binoculars_Prop",
+            "Binoculars", "Bugle_Prop Variant", "Bugle", "Compass", "Fannypack", "Glider",
+            "Jetpack", "Parasol", "Parasol_Roots Variant", "Parachute", "Pirate Compass",
+            "Rocketpack");
         Add(tags, ItemFilterTag.Deployable,
-            "Anti-Rope Cannon", "Anti-Rope Spool", "Bounce Fungus", "Chain Launcher",
-            "Checkpoint Flag", "Cloud Fungus", "Magic Bean", "Piton", "Portable Stove",
-            "Rope Cannon", "Rope Spool", "Scout Cannon", "Shelf Fungus");
+            "RopeShooterAnti", "Anti-Rope Spool", "BounceShroom", "ChainShooter",
+            "Flag_Plantable_Checkpoint", "CloudFungus", "MagicBean", "ClimbingSpike",
+            "PortableStovetopItem", "RopeShooter", "RopeSpool", "ScoutCannonItem",
+            "ShelfShroom");
         Add(tags, ItemFilterTag.Mystical,
-            "Ancient Idol", "Anti-Rope Cannon", "Anti-Rope Spool", "Anti-Zooka", "Cure-All",
-            "Cursed Skull", "Faerie Lantern", "Pandora's Lunchbox", "Ritual Dagger",
-            "Scout Effigy", "Scout's Ambition", "Scout's Generosity", "Scout's Honor",
-            "Scout's Initiative", "Scout's Tenacity", "Scoutmaster's Bugle", "Strange Gem",
-            "The Book of Bones", "Warp Compass");
+            "AncientIdol", "RopeShooterAnti", "Anti-Rope Spool", "AntiZooka", "Cure-All",
+            "Cursed Skull", "Lantern_Faerie", "PandorasBox", "RitualDagger", "ScoutEffigy",
+            "Amulet_InfiniteStamina", "Amulet_Clone", "ScoutsHonor", "Amulet_SuperJump",
+            "Amulet_Healing", "Bugle_Scoutmaster Variant", "Strange Gem", "BookOfBones",
+            "Warp Compass", "Cheat Compass", "Cheat Compass 1");
         return tags;
     }